services/_template/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.12.4
pydantic-settings==2.11.0

# HTTP Client (for downstream calls)
httpx==0.28.1

# JWT Authentication
PyJWT==2.10.1

# Database (async)
asyncpg==0.31.0
alembic==1.14.1
sqlalchemy[asyncio]==2.0.41

# Redis (caching, session)
redis[hiredis]==5.2.1

# Observability
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-exporter-otlp==1.34.0
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
